Add `load_fixture_flat` function.
`load_fixture_flat` is an alternative loader that accepts
`__key__` and `__parent__` to model entities with ancestry/parents,
as opposed to the nested structure that `load_fixture` expects.

To model entities with reference properties to keys (i.e.
`KeyProperty`), instead of specifying the nested entity, one can
specify a representation of a key/id instead. e.g.:

    "some_key_property__key__": ["Kind", "foo"]

or

    "some_key_property__id__": "foo"

For the second example, the entity kind will be determined by the
model's class definition.

See `tests/purchases_key.json` and `tests/purchases_parent.json` for
examples.